# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clinic Digital Growth SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new doctor/clinic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Clinics
  /clinics:
    get:
      operationId: listClinics
      tags: [clinics]
      summary: List all clinics (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Clinic"

    post:
      operationId: createClinic
      tags: [clinics]
      summary: Create a clinic
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClinicBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clinics/{clinicId}:
    get:
      operationId: getClinic
      tags: [clinics]
      summary: Get a clinic by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/clinicId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateClinic
      tags: [clinics]
      summary: Update a clinic
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/clinicId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClinicBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clinics/{clinicId}/suspend:
    post:
      operationId: suspendClinic
      tags: [clinics]
      summary: Suspend or reactivate a clinic (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/clinicId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuspendClinicBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"

  # Doctors
...
  /appointments/book:
    post:
      operationId: bookAppointment
      tags: [appointments]
      summary: Book an appointment (public - no auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookAppointmentBody"
      responses:
        "201":
          description: Booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingConfirmation"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments for the authenticated doctor's clinic
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
          description: Filter by date (YYYY-MM-DD)
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status
        - name: patientName
          in: query
          schema:
            type: string
          description: Filter by patient name
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentWithToken"
  /appointments/{appointmentId}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get an appointment
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/appointmentId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithToken"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment status or reschedule
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/appointmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithToken"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Token Queue
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    clinicId:
      name: clinicId
      in: path
      required: true
      schema:
        type: integer
...
    appointmentId:
      name: appointmentId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Auth schemas
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        clinicName:
          type: string
        clinicAddress:
          type: string
        clinicCity:
          type: string
      required:
        - name
        - email
        - phone
        - password
        - clinicName
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [super_admin, doctor, patient]
        clinicId:
          type: ["integer", "null"]
        specialization:
          type: ["string", "null"]
        qualification:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - phone
        - role
        - createdAt

    # Clinic schemas
    Clinic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        pincode:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        googleMapsUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        isSuspended:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - ownerName
        - phone
        - email
        - isActive
        - isSuspended
        - createdAt
        - updatedAt
    CreateClinicBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        whatsappNumber:
          type: string
        googleMapsUrl:
          type: string
      required:
        - name
        - slug
        - ownerName
        - phone
        - email
    UpdateClinicBody:
      type: object
      properties:
        name:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        logoUrl:
          type: string
        whatsappNumber:
          type: string
        googleMapsUrl:
          type: string
    SuspendClinicBody:
      type: object
      properties:
        suspend:
          type: boolean
      required:
        - suspend
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        clinicId:
          type: integer
        doctorId:
          type: ["integer", "null"]
        patientName:
          type: string
        patientPhone:
          type: string
        patientEmail:
          type: ["string", "null"]
        appointmentDate:
          type: string
        timeSlot:
          type: string
        reason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [waiting, in_progress, completed, cancelled, rescheduled, no_show]
        rescheduledDate:
          type: ["string", "null"]
        rescheduledTimeSlot:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clinicId
        - patientName
        - patientPhone
        - appointmentDate
        - timeSlot
        - status
        - createdAt
        - updatedAt
    AppointmentWithToken:
      allOf:
        - $ref: "#/components/schemas/Appointment"
        - type: object
          properties:
            token:
              $ref: "#/components/schemas/Token"
    BookAppointmentBody:
      type: object
      properties:
        clinicId:
          type: integer
        patientName:
          type: string
        patientPhone:
          type: string
        patientEmail:
          type: string
        appointmentDate:
          type: string
        timeSlot:
          type: string
        reason:
          type: string
      required:
        - clinicId
        - patientName
        - patientPhone
        - appointmentDate
        - timeSlot
    BookingConfirmation:
      type: object
      properties:
        appointment:
          $ref: "#/components/schemas/Appointment"
        token:
          $ref: "#/components/schemas/Token"
        message:
          type: string
      required:
        - appointment
        - token
        - message
    UpdateAppointmentBody:
      type: object
      properties:
        status:
          type: string
          enum: [waiting, in_progress, completed, cancelled, rescheduled, no_show]
        rescheduledDate:
          type: string
        rescheduledTimeSlot:
          type: string
        notes:
          type: string

    # Token schemas
    Token:
      type: object
      properties:
        id:
          type: integer
        clinicId:
          type: integer
        appointmentId:
          type: integer
        tokenNumber:
          type: integer
        tokenDate:
          type: string
        status:
          type: string
          enum: [waiting, called, completed, skipped]
        calledAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clinicId
        - appointmentId
        - tokenNumber
        - tokenDate
        - status
        - createdAt
...