# lib/api-zod/src/generated/types/bookAppointmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinic Digital Growth SaaS Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BookAppointmentBody {
  clinicId: number;
  patientName: string;
  patientPhone: string;
  patientEmail?: string;
  appointmentDate: string;
  timeSlot: string;
  reason?: string;
}